move chain unit tests to shared_state
